refactor: 代码模块化
